Add typemeta to gw resources in api release
