+++ internal/controller/apibinding/render/httproute.go
@@ -9,13 +9,11 @@ import (
 	"regexp"
 	"strings"
 
-	egv1a1 "github.com/envoyproxy/gateway/api/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/utils/ptr"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
-	gwapiv1a2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
 	choreov1 "github.com/openchoreo/openchoreo/api/v1"
 	dpkubernetes "github.com/openchoreo/openchoreo/internal/dataplane/kubernetes"
@@ -74,103 +72,6 @@ func makeRESTHTTPRoutes(rCtx *Context) []*choreov1.Resource {
 	return resources
 }
 
-// HTTPRouteFilters renders the HTTPRouteFilter resources for regex-based path replacement.
-func HTTPRouteFilters(rCtx *Context) []*choreov1.Resource {
-	apiType := rCtx.API.Spec.Type
-	switch apiType {
-	case choreov1.EndpointTypeREST:
-		return makeRESTHTTPRouteFilters(rCtx)
-	default:
-		rCtx.AddError(fmt.Errorf("unsupported API type: %s", apiType))
-		return nil
-	}
-}
-
-func makeRESTHTTPRouteFilters(rCtx *Context) []*choreov1.Resource {
-	if rCtx.API.Spec.RESTEndpoint == nil {
-		rCtx.AddError(fmt.Errorf("REST endpoint specification is missing"))
-		return nil
-	}
-	if rCtx.APIClass.Spec.RESTPolicy == nil {
-		rCtx.AddError(fmt.Errorf("REST policy is not defined in the API class"))
-		return nil
-	}
-
-	// Generate HTTPRouteFilters for each expose level and operation
-	httpRouteFilters := make([]*egv1a1.HTTPRouteFilter, 0)
-
-	// Process each operation and its expose levels
-	for _, operation := range rCtx.API.Spec.RESTEndpoint.Operations {
-		for _, exposeLevel := range operation.ExposeLevels {
-			if exposeLevel == choreov1.ExposeLevelProject {
-				continue // Skip project level for now
-			}
-			httpRouteFilter := makeHTTPRouteFilterForRestOperation(rCtx, operation, exposeLevel)
-			if httpRouteFilter != nil {
-				httpRouteFilters = append(httpRouteFilters, httpRouteFilter)
-			}
-		}
-	}
-
-	resources := make([]*choreov1.Resource, 0, len(httpRouteFilters))
-
-	for _, httpRouteFilter := range httpRouteFilters {
-		rawExt := &runtime.RawExtension{}
-		rawExt.Object = httpRouteFilter
-
-		resources = append(resources, &choreov1.Resource{
-			ID:     makeHTTPRouteFilterResourceID(httpRouteFilter),
-			Object: rawExt,
-		})
-	}
-
-	return resources
-}
-
-func makeHTTPRouteFilterForRestOperation(rCtx *Context, restOperation choreov1.RESTEndpointOperation,
-	exposeLevel choreov1.RESTOperationExposeLevel) *egv1a1.HTTPRouteFilter {
-	basePath := rCtx.API.Spec.RESTEndpoint.Backend.BasePath
-	endpointPath := path.Join(basePath, restOperation.Path)
-	pattern := GenerateRegexWithCaptureGroup(basePath, restOperation.Path, endpointPath)
-
-	filter := &egv1a1.HTTPRouteFilter{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      makeHTTPRouteName(rCtx, restOperation, exposeLevel),
-			Namespace: makeNamespaceName(rCtx),
-			Labels:    makeAPILabels(rCtx),
-		},
-		Spec: egv1a1.HTTPRouteFilterSpec{
-			URLRewrite: &egv1a1.HTTPURLRewriteFilter{
-				Path: &egv1a1.HTTPPathModifier{
-					Type: egv1a1.RegexHTTPPathModifier,
-					ReplaceRegexMatch: &egv1a1.ReplaceRegexMatch{
-						Pattern:      pattern,
-						Substitution: "\\1",
-					},
-				},
-			},
-		},
-	}
-
-	return filter
-}
-
-func makeHTTPRouteFilterResourceID(httpRouteFilter *egv1a1.HTTPRouteFilter) string {
-	return httpRouteFilter.Name
-}
-
-// SecurityPolicies renders the SecurityPolicy resources for the API.
-func SecurityPolicies(rCtx *Context) []*choreov1.Resource {
-	apiType := rCtx.API.Spec.Type
-	switch apiType {
-	case choreov1.EndpointTypeREST:
-		return makeSecurityPolicies(rCtx)
-	default:
-		rCtx.AddError(fmt.Errorf("unsupported API type: %s", apiType))
-		return nil
-	}
-}
-
 func makeHTTPRouteForRestOperation(rCtx *Context, restOperation choreov1.RESTEndpointOperation,
 	exposeLevel choreov1.RESTOperationExposeLevel) *gwapiv1.HTTPRoute {
 	pathType := gwapiv1.PathMatchRegularExpression
@@ -184,6 +85,10 @@ func makeHTTPRouteForRestOperation(rCtx *Context, restOperation choreov1.RESTEnd
 	regexEpPath := GenerateRegexWithCaptureGroup(basePath, restOperation.Path, endpointPath)
 
 	return &gwapiv1.HTTPRoute{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "gateway.networking.k8s.io/v1",
+			Kind:       "HTTPRoute",
+		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      name,
 			Namespace: makeNamespaceName(rCtx),
@@ -298,133 +203,6 @@ func getGatewayName(rCtx *Context) string {
 
 	// If no public operations found, return internal gateway
 	return defaultGateway
-}
-
-func makeSecurityPolicies(rCtx *Context) []*choreov1.Resource {
-	if rCtx.API.Spec.RESTEndpoint == nil {
-		rCtx.AddError(fmt.Errorf("REST endpoint specification is missing"))
-		return nil
-	}
-	if rCtx.APIClass.Spec.RESTPolicy == nil {
-		rCtx.AddError(fmt.Errorf("REST policy is not defined in the API class"))
-		return nil
-	}
-
-	// Generate SecurityPolicy for each expose level and operation
-	securityPolicies := make([]*egv1a1.SecurityPolicy, 0)
-
-	// Process each operation and its expose levels
-	for _, operation := range rCtx.API.Spec.RESTEndpoint.Operations {
-		for _, exposeLevel := range operation.ExposeLevels {
-			if exposeLevel == choreov1.ExposeLevelProject {
-				continue // Skip project level for now
-			}
-
-			// Get merged policies for this expose level
-			mergedPolicy, err := MergePoliciesForExposeLevel(rCtx.APIClass.Spec.RESTPolicy, exposeLevel)
-			if err != nil {
-				rCtx.AddError(fmt.Errorf("failed to merge policies for expose level %s: %w", exposeLevel, err))
-				continue
-			}
-
-			securityPolicy := makeSecurityPolicyForRestOperation(rCtx, operation, exposeLevel, mergedPolicy)
-			if securityPolicy != nil {
-				securityPolicies = append(securityPolicies, securityPolicy)
-			}
-		}
-	}
-
-	resources := make([]*choreov1.Resource, 0, len(securityPolicies))
-
-	for _, securityPolicy := range securityPolicies {
-		rawExt := &runtime.RawExtension{}
-		rawExt.Object = securityPolicy
-
-		resources = append(resources, &choreov1.Resource{
-			ID:     makeSecurityPolicyResourceID(securityPolicy),
-			Object: rawExt,
-		})
-	}
-
-	return resources
-}
-
-func makeSecurityPolicyForRestOperation(rCtx *Context, restOperation choreov1.RESTEndpointOperation,
-	exposeLevel choreov1.RESTOperationExposeLevel, mergedPolicy *choreov1.RESTPolicyWithConditionals) *egv1a1.SecurityPolicy {
-	name := makeHTTPRouteName(rCtx, restOperation, exposeLevel)
-	actionDeny := egv1a1.AuthorizationActionDeny
-	actionAllow := egv1a1.AuthorizationActionAllow
-
-	// Convert RESTOperation.Scopes to []egv1a1.JWTScope
-	jwtScopes := make([]egv1a1.JWTScope, len(restOperation.Scopes))
-	for i, scope := range restOperation.Scopes {
-		jwtScopes[i] = egv1a1.JWTScope(scope)
-	}
-
-	securityPolicy := &egv1a1.SecurityPolicy{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: makeNamespaceName(rCtx),
-			Labels:    makeAPILabels(rCtx),
-		},
-		Spec: egv1a1.SecurityPolicySpec{
-			PolicyTargetReferences: egv1a1.PolicyTargetReferences{
-				TargetRefs: []gwapiv1a2.LocalPolicyTargetReferenceWithSectionName{
-					{
-						LocalPolicyTargetReference: gwapiv1a2.LocalPolicyTargetReference{
-							Group: gwapiv1.GroupName,
-							Kind:  "HTTPRoute",
-							Name:  gwapiv1a2.ObjectName(name),
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Configure authentication if available
-	if mergedPolicy != nil && mergedPolicy.Authentication != nil {
-		if mergedPolicy.Authentication.Type == "jwt" {
-			if mergedPolicy.Authentication.JWT != nil {
-				securityPolicy.Spec.JWT = &egv1a1.JWT{
-					Providers: []egv1a1.JWTProvider{
-						{
-							Name: "default",
-							RemoteJWKS: egv1a1.RemoteJWKS{
-								URI: mergedPolicy.Authentication.JWT.JWKS,
-							},
-							Issuer:    mergedPolicy.Authentication.JWT.Issuer,
-							Audiences: mergedPolicy.Authentication.JWT.Audience,
-						},
-					},
-				}
-
-				// Add authorization if scopes are present
-				if len(jwtScopes) > 0 {
-					securityPolicy.Spec.Authorization = &egv1a1.Authorization{
-						Rules: []egv1a1.AuthorizationRule{
-							{
-								Principal: egv1a1.Principal{
-									JWT: &egv1a1.JWTPrincipal{
-										Provider: "default",
-										Scopes:   jwtScopes,
-									},
-								},
-								Action: actionAllow,
-							},
-						},
-						DefaultAction: &actionDeny,
-					}
-				}
-			}
-		}
-	}
-
-	return securityPolicy
-}
-
-func makeSecurityPolicyResourceID(policy *egv1a1.SecurityPolicy) string {
-	return policy.Name
 }
 
 func makeHTTPRouteResourceID(httpRoute *gwapiv1.HTTPRoute) string {
